[WIP][IR] Continue Tensor PrimType boundary fixes

+++ src/s_tir/analysis/calculate_allocated_memory.cc
@@ -76,7 +76,7 @@ class AllocBufferCalculator : public StmtExprVisitor {
         break;
       }
     }
-    size *= (((op->buffer->dtype).bits + 7) / 8) * static_cast<int16_t>((op->buffer->dtype).lanes);
+    size *= ((op->buffer->dtype.bits() + 7) / 8) * op->buffer->dtype.lanes();
     _current_size[storage_scope] += size;
     _max_size[storage_scope] = std::max(_current_size[storage_scope], _max_size[storage_scope]);
     StmtExprVisitor::VisitStmt_(op);
